refactor: migrate from Spring Data JPA/MongoDB to JDBC/MongoDB Driver with custom DAO layer
Replace Spring Data JPA and Spring Data MongoDB repositories with custom DAO implementations using JdbcTemplate and MongoTemplate for direct database access. This migration provides more control over query execution and reduces framework abstraction overhead.

Key changes:
- Remove Spring Data JPA/MongoDB repository dependencies
- Implement custom DAO layer with JdbcTemplate for PostgreSQL and MongoTemplate for MongoDB
- Update all services to use new DAO interfaces instead of repositories
- Refactor pagination to use custom implementation instead of Spring Data Pageable
- Add comprehensive database configuration for both PostgreSQL and MongoDB
- Update cache layer to work with new DAO implementations
- Add detailed documentation for caching strategy and testing approach
- Enhance cache metrics tracking with hit/miss/eviction statistics
- Fix type conversions between Integer and Long IDs across the application

+++ blog-backend/src/main/java/com/kratosgado/blog/backend/controllers/AuthController.java
@@ -41,7 +41,7 @@ public class AuthController {
 
     return new AuthResponse(
         token,
-        user.getId(),
+        user.getId().longValue(),
         user.getUsername(),
         user.getEmail(),
         user.getRole());
@@ -61,7 +61,7 @@ public class AuthController {
 
     return new AuthResponse(
         token,
-        user.getId(),
+        user.getId().longValue(),
         user.getUsername(),
         user.getEmail(),
         user.getRole());
